Refactor: scheduler: move shared clone/bundle code from pcmk_sched_clone.c
... into a new source file pcmk_sched_instances.c, since it applies to both
clone instances and bundle replica containers.

No code changes, except that distribute_children() is now declared in the
library-private header instead of each file that uses it.
